refactor: remove barrel export, use direct imports
- Delete schemas/index.ts barrel file
- Update all imports to use direct paths to source files
