+++ packages/twenty-server/src/modules/workflow/workflow-tools/tools/record-step-configurators/delete-record-step-configurator.factory.ts
@@ -1,7 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
 import { z } from 'zod';
 
-import { type ObjectMetadataForToolSchema } from 'src/engine/core-modules/record-crud/schemas';
+import { type ObjectMetadataForToolSchema } from 'src/engine/core-modules/record-crud/types/object-metadata-for-tool-schema.type';
 import { WorkflowActionType } from 'src/modules/workflow/workflow-executor/workflow-actions/types/workflow-action.type';
 import { type WorkflowToolContext } from 'src/modules/workflow/workflow-tools/types/workflow-tool-dependencies.type';
